+++ packages/react/src/components/special-components/HeaderBottomComponent.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { useSpecialComponentProps } from '../../hooks/useSpecialComponentProps';
+
+export function HeaderBottomComponent() {
+  const { props } = useSpecialComponentProps();
+  const Component = props.config.specialComponents?.headerBottom;
+
+  if (!Component) return null;
+
+  return <Component {...props} />;
+}