+++ packages/react-headless/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opencx/widget-react-headless
 
+## 4.0.30
+
+### Patch Changes
+
+- add `WidgetConfig.specialComponents.headerBottom`
+- Updated dependencies
+  - @opencx/widget-core@4.0.30
+
 ## 4.0.29
 
 ### Patch Changes